[test] gathering pk tests
