+++ tests/integration/nano/test_public_key.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
+# Copyright 2024 Functori <contact@functori.com>
+# Copyright 2024 Trilitech <contact@trili.tech>
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,16 +14,41 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Module to test get-public-key instruction."""
-
-from pathlib import Path
+"""Gathering of tests related to public key."""
 
 from utils.account import Account, SigType
-from utils.app import Screen, TezosAppScreen
+from utils.app import Screen, TezosAppScreen, DEFAULT_ACCOUNT
+
+def test_get_pk(app: TezosAppScreen):
+    """Test that public keys get from the app are correct."""
+    accounts = [
+        Account("m/44'/1729'/0'/0'",
+                SigType.ED25519,
+                "edpkuXX2VdkdXzkN11oLCb8Aurdo1BTAtQiK8ZY9UPj2YMt3AHEpcY"),
+        Account("m/44'/1729'/0'/0'",
+                SigType.SECP256K1,
+                "sppk7bVy617DmGvXsMqcwsiLtnedTN2trUi5ugXcNig7en4rHJyunK1"),
+        Account("m/44'/1729'/0'/0'",
+                SigType.SECP256R1,
+                "p2pk67fq5pzuMMABZ9RDrooYbLrgmnQbLt8z7PTGM9mskf7LXS5tdBG"),
+        Account("m/44'/1729'/0'/0'",
+                SigType.BIP32_ED25519,
+                "edpkumJgSsSxkpiB5hmTq6eZcrmc6BsJtLAhYceFTiziFqje4mongz")
+    ]
+
+    for account in accounts:
+
+        app.assert_screen(Screen.HOME)
+
+        data = app.backend.get_public_key(account, with_prompt=False)
+
+        app.checker.check_public_key(account, data)
+
+    app.quit()
 
 def test_provide_pk(app: TezosAppScreen):
     """Test that public keys get from the app are correct and correctly displayed."""
-    test_name = Path(__file__).stem
+    test_name = "test_provide_pk"
 
     app.assert_screen(Screen.HOME)
 
@@ -60,5 +86,15 @@ def test_provide_pk(app: TezosAppScreen):
         data = app.provide_public_key(account, test_name + "/" + kind)
 
         app.checker.check_public_key(account, data)
+
+    app.quit()
+
+def test_reject_pk(app: TezosAppScreen):
+    """Check reject pk behaviour"""
+    test_name = "test_reject_pk"
+
+    app.assert_screen(Screen.HOME)
+
+    app.reject_public_key(DEFAULT_ACCOUNT, test_name)
 
     app.quit()
